fix in metadata event type read funtion
Summary:
Fix to avoid signed extesnion when the value_length is a large number (such
that the MSB is set). This happens in a config change event when the
length of the config change string is large (causing the MSB in the
16bit unsigned integer to be set). While reading this back we have to
ensure the signed conversion is done correctly.

For example, a config change string of length 45460 bytes (0xb194) gets
sign extended to 65428 bytes (0xff94)

Reviewed By: anirbanr-fb

Differential Revision: D26496395

+++ libbinlogevents/src/control_events.cpp
@@ -424,7 +424,7 @@ Metadata_event::Metadata_event(const char *buf,
   READER_TRY_INITIALIZATION;
   READER_ASSERT_POSITION(fde->common_header_len);
 
-  uint8_t read_len [[maybe_unused]] = fde->common_header_len;
+  uint32_t read_len [[maybe_unused]] = fde->common_header_len;
 
   /* Read and intialize every type in the stream for this event */
   while (READER_CALL(available_to_read) > 0) {
@@ -554,7 +554,7 @@ uint Metadata_event::read_type(Metadata_event_types type) {
   using MET = Metadata_event_types;
 
   // Read the 'length' of the field's value
-  uint value_length = 0;
+  uint16_t value_length = 0;
   uint64_t hlc_time = 0;
   uint64_t prev_hlc_time = 0;
   int64_t term = -1, index = -1;
